+++ transcustomfilter.h
@@ -7,12 +7,12 @@
 class TransCustomFilter : public Transformation
 {
 public:
-    TransCustomFilter(const std::vector<int> &params);
-    TransCustomFilter(int left, int top, int right, int bottom, const std::vector<int> &params);
+    TransCustomFilter(const std::vector<float> &params);
+    TransCustomFilter(int left, int top, int right, int bottom, const std::vector<float> &params);
     QString toString() const;
     virtual TransCustomFilter* clone() const;
 private:
-    std::vector<int> mask;
+    std::vector<float> mask;
 };
 
 #endif // TRANSCUSTOMFILTER_H
